Pushing first pass of updates
